fixed Issue 282 : third try
git-svn-id: http://tint2.googlecode.com/svn/trunk@528 121b4492-b84c-0410-8b4c-0d4edfb3f3cc

+++ src/taskbar/task.c
@@ -209,7 +209,7 @@ void get_icon (Task *tsk)
 	int i;
 	Imlib_Image img = NULL;
 	XWMHints *hints = 0;
-	long *data = 0;
+	int32_t *data = 0;
 
 	int k;
 	for (k=0; k<TASK_STATE_COUNT; ++k) {
@@ -224,7 +224,7 @@ void get_icon (Task *tsk)
 	if (data) {
 		// get ARGB icon
 		int w, h;
-		long *tmp_data;
+		int32_t *tmp_data;
 
 		tmp_data = get_best_icon (data, get_icon_count (data, i), i, &w, &h, panel->g_task.icon_size1);
